Switched Celestial to static library, fixed all warning (sans spdlog-related), Created generic and OpenGL contexts, Drawn first triangle.

+++ Celestial/src/Celestial/Events/Event.h
@@ -61,7 +61,7 @@ namespace Celestial
 
 		inline bool IsInCategory(EventCategory category)
 		{
-			return GetCategoryFlags() & category;
+			return GetCategoryFlags() & static_cast<int>(category);
 		}
 	public:
 		bool Handled = false;
